% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaqueteCeibo.R
\name{Tabla5}
\alias{Tabla5}
\title{Una función para datos de Ceibo}
\usage{
Tabla5(x)
}
\arguments{
\item{Longitud}{de vastago}
}
\description{
Esta función ayuda a procesar los datos tomados por semana.
}
\examples{
Este ejemplo es para la semana 5, que sería Tabla5
}
\keyword{Ceibo}
